Model updated. Appearance of the views changed. Ticker View started.

+++ Gainy/Gainy/Model/CollectionDetails/TickerFinancialMetrics/TickerFinancialMetrics.swift
@@ -1,8 +1,8 @@
 struct TickerFinancialMetrics {
     let todaysPriceChange: Float
     let currentPrice: Float
-    let dividendGrowthPercent: Int
+    let dividendGrowthPercent: Float
     let priceToEarnings: Float
-    let marketCapitalization: Int64
+    let marketCapitalization: Float
     let highlight: String
 }
